# gateway/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Описание эндпоинтов сервиса анонимизации данных",
        "title": "Anonix Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/kinds": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список видов данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kinds"
                ],
                "summary": "Получить список видов данных",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.KindSchema"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to get kind list"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый вид данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kinds"
                ],
                "summary": "Создать вид данных",
                "parameters": [
                    {
                        "description": "Данные вида данных",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateKindSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.KindSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request"
                    },
                    "409": {
                        "description": "kind already exists"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/kinds/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает объект вида данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kinds"
                ],
                "summary": "Получить вид данных по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вида данных",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.KindSchema"
                        }
                    },
                    "400": {
                        "description": "invalid kind ID"
                    },
                    "404": {
                        "description": "kind not found"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет вид данных по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kinds"
                ],
                "summary": "Удалить вид данных",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вида данных",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid kind ID"
                    },
                    "404": {
                        "description": "kind not found"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает обновленный вид данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kinds"
                ],
                "summary": "Обновить вид данных",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вида данных",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateKindSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.KindSchema"
                        }
                    },
                    "400": {
                        "description": "invalid kind ID"
                    },
                    "404": {
                        "description": "kind not found"
                    },
                    "500": {
                        "description": "internal error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schemas.CreateKindSchema": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 3
                },
                "mask": {
                    "type": "string",
                    "example": "^\\d{4} \\d{6}$"
                },
                "name": {
                    "type": "string",
                    "example": "passport"
                },
                "russian_name": {
                    "type": "string",
                    "example": "Паспорт"
                },
                "short_name": {
                    "type": "string",
                    "example": "psp"
                }
            }
        },
...
        "schemas.KindSchema": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 3
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mask": {
                    "type": "string",
                    "example": "^\\d{4} \\d{6}$"
                },
                "name": {
                    "type": "string",
                    "example": "passport"
                },
                "russian_name": {
                    "type": "string",
                    "example": "Паспорт"
                },
                "short_name": {
                    "type": "string",
                    "example": "psp"
                }
            }
        },
...
        "schemas.UpdateKindSchema": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 3
                },
                "mask": {
                    "type": "string",
                    "example": "^\\d{4} \\d{6}$"
                },
                "name": {
                    "type": "string",
                    "example": "passport"
                },
                "russian_name": {
                    "type": "string",
                    "example": "Паспорт"
                },
                "short_name": {
                    "type": "string",
                    "example": "psp"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}